pyproject.toml
[tool.poetry]
name = "patientrack-backend"
version = "0.1.0"
description = ""
authors = ["giancarloprieto"]
readme = "README.md"
packages = [{include = "patientrack_backend"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2.6"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
pillow = "^10.1.0"
django-storages = "^1.14.2"
boto3 = "^1.33.11"
django-rest-framework = "^0.1.0"
pandas = "1.3.3"
xlsxwriter = "^3.1.9"
zappa = "^0.58.0"
django-pwa = "^1.1.0"
whitenoise = "^6.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
